requirements.txt
py-trello==0.18.0
pyTelegramBotAPI==4.3.1
schedule==1.1.0